services/file-service/requirements.txt
Django==5.0.1
djangorestframework-simplejwt==5.3.1
django-ninja==1.1.0
celery==5.3.4
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.1
Pillow==10.2.0
requests==2.31.0


